requirements.txt
django
djoser
pillow
djangorestframework
django-cors-headers
django_rest_passwordreset